// packages/svelte-reveal/src/internal/types.ts
export interface RevealOptions {
  /**
   * When set to false, the transition for the target element is disabled.
   */
  disable?: boolean;
  /**
   * It enables/disables debugging mode for the target DOM element.
   * This will log to the console the target DOM element, along with the options and config.
   *
   * In order to be able to use this mode, you are required to also set the `ref` property.
   */
  debug?: boolean;
  /**
   * When `debug` is set to `true`, you are required to specificy a `ref` string.
   *
   * When multiple DOM nodes have debug mode enabled, `ref` strings allow you to
   * know to which DOM node a console log statement belongs to.
   */
  ref?: string;
  /**
   * When set to true, the console logs of the target node are colored,
   * making it easier to see them among many other logs.
   */
  highlightLogs?: boolean;
  /**
   * The color to use to color the console logs when the `highlightLogs` option is also set to true.
   *
   * Any valid CSS color can be used here.
   */
  highlightColor?: string;
  /**
   * The root element used by the Intersection Observer.
   */
  root?: IntersectionObserver['root'];
  /**
   * The root margin property of the Intersection Observer.
   */
  rootMargin?: IntersectionObserver['rootMargin'];
  /**
   * The threshold (in percentage from 0.0 to 1.0) property used by the Intersection
   * Observer to know when its target element is considered visible.
   */
  threshold?: number;
  /**
   * The type of transition that is triggered when the target node becomes visible.
   */
  transition?: Transition;
  /**
   * When set to true, the node transitions out when out of view, and is revealed again when back in view.
   *
   * ⚠️ Be careful not to overuse this option.
   */
  reset?: boolean;
  /**
   * How long the transition lasts (in milliseconds).
   */
  duration?: number;
  /**
   * How long the transition is delayed (in milliseconds) before being triggered.
   */
  delay?: number;
  /**
   * The type of easing function applied to the `transition`.
   */
  easing?: Easing;
  /**
   * The individual weights of a custom cubic-bezier curve.
   * This option is necessary when `easing` is set to `custom`.
   */
  customEasing?: CustomEasing;
  /**
   * The starting offset position in pixels on the x-axis of the `"slide"` transition.
   * If `x` is negative, the element will transition from the left, else from the right.
   */
  x?: number;
  /**
   * The starting offset position in pixels on the y-axis of the `"fly"` transition.
   * If `y` is negative, the element will transition from the top, else from the bottom.
   */
  y?: number;
  /**
   * The starting rotation offset in degrees of the `"spin"` transition.
   * If `rotate` is positive, the element will spin clockwise, else counter-clockwise.
   */
  rotate?: number;
  /**
   * The starting opacity value in percentage of any transition.
   * It can be a number between `0.0` and `1.0`.
   */
  opacity?: number;
  /**
   * The starting blur value in pixels of the `"blur"` transition.
   */
  blur?: number;
  /**
   * The starting scale value in percentage of the `"scale"` transition.
   */
  scale?: number;
  /**
   * Function that gets fired when the node starts being revealed.
   */
  onRevealStart?: (node: HTMLElement) => void;
  /**
   * Function that gets fired when the node is fully revealed.
   */
  onRevealEnd?: (node: HTMLElement) => void;
  /**
   * Function that gets fired when the `reset` option is set to `true`
   * and the node starts transitioning out.
   */
  onResetStart?: (node: HTMLElement) => void;
  /**
   * Function that gets fired when the `reset` option is set to `true`
   * and the node has fully transitioned out.
   */
  onResetEnd?: (node: HTMLElement) => void;
  /**
   * Function that gets fired when the node is mounted on the DOM.
   */
  onMount?: (node: HTMLElement) => void;
  /**
   * Function that gets fired when the action options are updated.
   */
  onUpdate?: (node: HTMLElement) => void;
  /**
   * Function that gets fired when the node is unmounted from the DOM.
   */
  onDestroy?: (node: HTMLElement) => void;
}
// ...
export type Transition = 'fly' | 'fade' | 'blur' | 'scale' | 'slide' | 'spin';

/**
 * The types of supported easing functions that can be used to tweak the timing of a transition.
 */
export type Easing =
  | 'linear'
  | 'easeInSine'
  | 'easeOutSine'
  | 'easeInOutSine'
  | 'easeInQuad'
  | 'easeOutQuad'
  | 'easeInOutQuad'
  | 'easeInCubic'
  | 'easeOutCubic'
  | 'easeInOutCubic'
  | 'easeInQuart'
  | 'easeOutQuart'
  | 'easeInOutQuart'
  | 'easeInQuint'
  | 'easeOutQuint'
  | 'easeInOutQuint'
  | 'easeInExpo'
  | 'easeOutExpo'
  | 'easeInOutExpo'
  | 'easeInCirc'
  | 'easeOutCirc'
  | 'easeInOutCirc'
  | 'easeInBack'
  | 'easeOutBack'
  | 'easeInOutBack'
  | 'custom';

/**
 * The individual weights of a custom cubic-bezier curve.
 */
export type CustomEasing = [number, number, number, number];